test: move new tracker tests in packages matching implementation
